[UPSTREAM] tests/sys/audit: XFAIL extattr tests on tmpfs mounts
In the CheriBSD CI, we run the testsuite with /tmp as tmpfs. This causes
the extattr audit tests to fail since tmpfs does not (yet) support
extattrs. XFAIL those tests if the target path is on a tmpfs mount.

While touching these two files also convert the ATF_REQUIRE_EQ(-1, checks
to use ATF_REQURIE_ERRNO().