purecap-kernel: Fix for sockaddr_dl out-of-bounds access.
In some cases, sockaddr_dl instances allocated on the stack are cast
to struct sockaddr * and make their way to rtsock_msg_mbuf(), which
copies the padded SA_SIZE to an mbuf.
The padding is never guaranteed to be allocated for these sockaddr_dl,
nor for any other that does not properly round the size prior to malloc().
One instance of this has been triggered in in6_newaddrmsg() by the gateway
sockaddr_dl struct. This causes the copy to leak 2 bytes of uninitialized
kernel stack to the mbuf, or a length violation with CHERI subobject-bounds.

To fix this we only copy sa_len bytes in rtsock_msg_mbuf() as this is the
only place where this issue appears, however the target buffer must
still include the padding as we expect the alignment on the receiver side.